feat: add UsuarioController with CRUD operations for user management
