+++ TheWalkingPets.Service/BLL/Services/Contract/IUsuarioService.cs
@@ -7,7 +7,7 @@ namespace TheWalkingPets.Service.BLL.Services.Contract
     {
         Task<Result<IEnumerable<UsuarioReadDto>>> GetAllAsync();
         Task<Result<UsuarioReadDto>> GetByIdAsync(Guid id);
-        Task<Result<UsuarioReadDto>> CreateAsync(UsuarioReadDto usuarioCreateDto);
+        Task<Result<UsuarioReadDto>> CreateAsync(UsuarioWriteDto usuarioCreateDto);
         Task<Result<UsuarioReadDto>> UpdateAsync(Guid id, UsuarioWriteDto usuarioUpdateDto);
         Task<Result> DeleteAsync(Guid id);
     }
