+++ TheWalkingPets.Service/Controllers/Extensions/ResultExtensions.cs
@@ -1,43 +1,43 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System.Runtime.CompilerServices;
 using TheWalkingPets.Service.common;
 
 namespace TheWalkingPets.Service.Controllers.Extensions
 {
-    public static class ResultExtencions
+    public static class ResultExtensions
     {
         public static IResult ToProblemDetailsResult(this Result result)
         {
             if (result.IsSuccess)
             {
-                throw new InvalidOperationException("cannot create problem details for a successful result. ");
+                throw new InvalidOperationException("Cannot create problem details for a successful result.");
             }
 
             var serverError = result.Error.Code.Contains("Unhandled");
 
             return Results.Problem(
                 statusCode: serverError ? StatusCodes.Status500InternalServerError : StatusCodes.Status400BadRequest,
-                title: serverError ? "serverError" : "Bad request",
-                type: serverError ? "https://datatracker.ietf.org/doc/html/rfc7231#section-6.6.1" : "https://tools.ietf.org/html/rfc7231#section-6.5.1",
+                title: serverError ? "Server Error" : "Bad Request",
+                 type: serverError ? "https://datatracker.ietf.org/doc/html/rfc7231#section-6.6.1" : "https://tools.ietf.org/html/rfc7231#section-6.5.1",
                 extensions: new Dictionary<string, object?>
                 {
-                    {"errors", new[] {result.Error} }
+                    { "errors", new[] {result.Error} }
                 }
             );
         }
+
         public static ActionResult ToProblemDetails(this Result result)
         {
-            if(result.IsSuccess)
+            if (result.IsSuccess)
             {
                 throw new InvalidOperationException("Cannot create problem details for a successful result.");
             }
 
-            var serverError = result.Error.Code.Contains("unhandled");
+            var serverError = result.Error.Code.Contains("Unhandled");
 
             var problemDetails = new ProblemDetails
             {
                 Status = serverError ? StatusCodes.Status500InternalServerError : StatusCodes.Status400BadRequest,
-                Title = serverError ? "sever Error" : "badRequest",
+                Title = serverError ? "Server Error" : "Bad Request",
                 Type = serverError ? "https://datatracker.ietf.org/doc/html/rfc7231#section-6.6.1" : "https://tools.ietf.org/html/rfc7231#section-6.5.1",
                 Extensions =
                 {
@@ -47,9 +47,8 @@ namespace TheWalkingPets.Service.Controllers.Extensions
 
             return new ObjectResult(problemDetails)
             {
-                StatusCode = serverError ? StatusCodes.Status500InternalServerError : StatusCodes.Status400BadRequest,
+                StatusCode = serverError ? StatusCodes.Status500InternalServerError : StatusCodes.Status400BadRequest
             };
         }
-
     }
 }
